requirements.txt
langchain
langchain-core
langchain-text-splitters
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
langchain_community
langchain_google_genai
dotenv
google-generativeai